Initial bug fixes
I was using the wrong stratification in previous commits, and I'm not sure the popflag was implemented correctly. My previous version assumed individuals were ordered in the gtypes object the same as in the original strata file, which is unlikely. Now it seems to be working correctly, and matching previous results