Complete and summarize analyses
I finished the runs Ttru_Run7_subset and Ttru_Run8_subset. I also ran a few tests to see what happens when I set popflag = 0 for soem randomly chosen Ttru instead of the CNMI animals. The results are pretty similar to when I run it for the CNMI animals, so apparently STRUCTURE just tends to assign animals as hybrids when you don't use a prior? Anyway, I also wrote a script to summarize the results for a Table in the paper.